+++ src/Entity/IdBasedEntity.php
@@ -1,0 +1,38 @@
+<?php
+
+/**
+ * This file is part of the EightMarq Symfony bundles.
+ *
+ * (c) Norbert Schvoy <norbert.schvoy@eightmarq.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace EightMarq\CoreBundle\Entity;
+
+use Doctrine\ORM\Mapping as ORM;
+use EightMarq\CoreBundle\Entity\Interfaces\IdBasedEntityInterface;
+use EightMarq\CoreBundle\Entity\Traits\IdBasedEntityMethodsTrait;
+use Knp\DoctrineBehaviors\Contract\Entity\BlameableInterface;
+use Knp\DoctrineBehaviors\Contract\Entity\SoftDeletableInterface;
+use Knp\DoctrineBehaviors\Contract\Entity\TimestampableInterface;
+use Knp\DoctrineBehaviors\Model\Blameable\BlameableTrait;
+use Knp\DoctrineBehaviors\Model\SoftDeletable\SoftDeletableTrait;
+use Knp\DoctrineBehaviors\Model\Timestampable\TimestampableTrait;
+
+#[ORM\MappedSuperclass]
+class IdBasedEntity implements IdBasedEntityInterface, TimestampableInterface, BlameableInterface, SoftDeletableInterface
+{
+    use IdBasedEntityMethodsTrait;
+    use BlameableTrait;
+    use SoftDeletableTrait;
+    use TimestampableTrait;
+
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column]
+    private int|null $id = null;
+}